Finished Refactor - Removed tests which were failing due to being unwritten. - Changed addGameObject() behavior in BasicGame to not register GameObjects as drawListeners. - Updated to Java18. - Removed Jacoco Gradle plugin as it was causing builds to fail.
